fix: replace the conformance bootstrap admin spec with an owner-email spec
seamless-auth-api removed the admin bootstrap invite flow, so the harness
call to POST /internal/bootstrap/admin-invite now 404s. The verify stack
sets OWNER_EMAIL instead, and the spec registers that address and asserts
the admin role is granted at signup.
